requirements.txt
Django==4.0.6
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.3
django-dotenv==1.4.2